refactor: 패키지 구조 변경 및 클래스 재배치
- 도메인 기준으로 패키지 구조 리팩토링
