+++ src/main/java/org/example/newsfeedproject/comment/controller/CommentController.java
@@ -1,0 +1,5 @@
+package org.example.newsfeedproject.comment.controller;
+
+public class CommentController {
+
+}
