+++ src/main/java/org/example/newsfeedproject/comment/dto/CommentCreateRequest.java
@@ -1,0 +1,10 @@
+package org.example.newsfeedproject.comment.dto;
+
+import jakarta.validation.constraints.NotNull;
+
+public class CommentCreateRequest {
+    @NotNull
+    private Long feedId;
+    @NotNull
+    private String comments;
+}
