+++ src/main/java/org/example/newsfeedproject/comment/repository/QCommentRepository.java
@@ -1,0 +1,4 @@
+package org.example.newsfeedproject.comment.repository;
+
+public interface QCommentRepository {
+}
