+++ src/main/java/org/example/newsfeedproject/comment/repository/QCommentRepositoryImpl.java
@@ -1,0 +1,4 @@
+package org.example.newsfeedproject.comment.repository;
+
+public class QCommentRepositoryImpl {
+}
