+++ src/main/java/org/example/newsfeedproject/commentlike/entity/CommentLike.java
@@ -1,0 +1,21 @@
+package org.example.newsfeedproject.commentlike.entity;
+
+import jakarta.persistence.*;
+import org.example.newsfeedproject.comment.entity.Comment;
+import org.example.newsfeedproject.domain.user.entity.User;
+
+@Entity
+@Table(name = "comment_like")
+public class CommentLike {
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id;
+
+    @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name = "user_id", nullable = false)
+    private User user;
+
+    @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name = "comment_id", nullable = false)
+    private Comment comment;
+}
