--- src/main/java/org/example/newsfeedproject/constant/Const.java
@@ -1,7 +1,0 @@
-package org.example.newsfeedproject.constant;
-
-public interface Const {
-
-    public static final String USER = "user";
-
-}
