--- src/main/java/org/example/newsfeedproject/domain/comment/CommentController.java
@@ -1,5 +1,0 @@
-package org.example.newsfeedproject.domain.comment;
-
-public class CommentController {
-
-}
