--- src/main/java/org/example/newsfeedproject/domain/comment/QCommentRepository.java
@@ -1,4 +1,0 @@
-package org.example.newsfeedproject.domain.comment;
-
-public interface QCommentRepository {
-}
