--- src/main/java/org/example/newsfeedproject/domain/comment/QCommentRepositoryImpl.java
@@ -1,4 +1,0 @@
-package org.example.newsfeedproject.domain.comment;
-
-public class QCommentRepositoryImpl {
-}
