--- src/main/java/org/example/newsfeedproject/domain/comment/dto/CommentCreateRequest.java
@@ -1,10 +1,0 @@
-package org.example.newsfeedproject.domain.comment.dto;
-
-import jakarta.validation.constraints.NotNull;
-
-public class CommentCreateRequest {
-    @NotNull
-    private Long feedId;
-    @NotNull
-    private String comments;
-}
