--- src/main/java/org/example/newsfeedproject/domain/common/entity/BaseEntity.java
@@ -1,24 +1,0 @@
-package org.example.newsfeedproject.domain.common.entity;
-
-import jakarta.persistence.Column;
-import jakarta.persistence.EntityListeners;
-import jakarta.persistence.MappedSuperclass;
-import lombok.Getter;
-import org.springframework.data.annotation.CreatedDate;
-import org.springframework.data.annotation.LastModifiedDate;
-import org.springframework.data.jpa.domain.support.AuditingEntityListener;
-
-import java.time.LocalDateTime;
-
-@Getter
-@MappedSuperclass
-@EntityListeners(AuditingEntityListener.class)
-public class BaseEntity {
-
-    @CreatedDate
-    @Column(updatable = false)
-    protected LocalDateTime createdAt;
-
-    @LastModifiedDate
-    protected LocalDateTime updatedAt;
-}
