--- src/main/java/org/example/newsfeedproject/domain/user/entity/User.java
@@ -1,28 +1,0 @@
-package org.example.newsfeedproject.domain.user.entity;
-
-import jakarta.persistence.*;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@Entity
-@Getter
-@NoArgsConstructor
-@Table(name = "user")
-public class User {
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    @Column(nullable = false)
-    private String email;
-
-    @Column(nullable = false)
-    private String nickname;
-
-    @Column(nullable = false)
-    private String password;
-
-    public User(Long id){
-        this.id = id;
-    }
-}
