+++ src/main/java/org/example/newsfeedproject/user/constant/Const.java
@@ -1,0 +1,7 @@
+package org.example.newsfeedproject.user.constant;
+
+public interface Const {
+
+    public static final String USER = "user";
+
+}
