+++ src/main/java/org/example/newsfeedproject/user/dto/LoginResponseDto.java
@@ -1,0 +1,17 @@
+package org.example.newsfeedproject.user.dto;
+
+import lombok.Getter;
+
+@Getter
+public class LoginResponseDto {
+
+    private final String nickname;
+    private final String email;
+    private final String message;
+
+    public LoginResponseDto(String nickname, String email, String message) {
+        this.nickname = nickname;
+        this.email = email;
+        this.message = message;
+    }
+}
