+++ src/main/java/org/example/newsfeedproject/user/entity/User.java
@@ -7,7 +7,7 @@ import jakarta.validation.constraints.Pattern;
 import jakarta.validation.constraints.Size;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
-import org.example.newsfeedproject.common.BaseTimeEntity;
+import org.example.newsfeedproject.common.entity.BaseTimeEntity;
 import org.hibernate.annotations.SQLDelete;
 import org.hibernate.annotations.SQLRestriction;
